Define website schema to fix build when db is empty
When all occurences of a field in the api reponse are empty, meaning the database has no entries , the field can not be inferred and will treated as non-existent by graphQL. This will make the build fail, as it relies on the queried fields to be present. To fix, we can manually define the schema and let graphQL know what fields to expect.